Added the Multiplication and division buttons Added functioning multiplication and division factors.

+++ JavaCalculator/src/Operations.java
@@ -1,15 +1,15 @@
 public class Operations {
-    public double addition(int x, int y) {
+    public int addition(int x, int y) {
         return x + y;
     }
-    public double subtraction(int x, int y) {
+    public int subtraction(int x, int y) {
         return x - y;
     }
-    public double multiplication(int x, int y) {
+    public int multiplication(int x, int y) {
         return x * y;
     }
-    public double division(int x, int y) {
-        return (double) x / y;
+    public int division(int x, int y) {
+        return  x / y;
     }
     public String addNumber(int num, String result) {
         result = result + num;
